requirements.txt
discord.py==1.3.4
python-dotenv==0.13.0
mysql-connector-python==8.0.21
asyncio==3.4.3
praw==6.5.1     
